refactor(types): consolidate wire-type files into domain groups
Fold the per-shape source and test files into profiles.go, roster.go, and
ranks.go (28 files down to 18). Pure relocation: no type, JSON tag, or test
changed; goimports recomputed the import blocks. Verified by an unchanged
test-function set and a green go build/vet/test.

> *This was generated by AI*
